+++ src/main/java/net/cravencraft/betterparagliders/mixins/paragliders/capabilities/ClientPlayerMovementMixin.java
@@ -1,11 +1,14 @@
 package net.cravencraft.betterparagliders.mixins.paragliders.capabilities;
 
+import net.bettercombat.api.MinecraftClient_BetterCombat;
 import net.bettercombat.logic.PlayerAttackProperties;
 import net.cravencraft.betterparagliders.network.ModNet;
 import net.cravencraft.betterparagliders.network.SyncActionToServerMsg;
+import net.minecraft.client.Minecraft;
 import net.minecraft.world.entity.player.Player;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
@@ -15,6 +18,7 @@ import tictim.paraglider.impl.movement.PlayerMovement;
 @Mixin(ClientPlayerMovement.class)
 public abstract class ClientPlayerMovementMixin extends PlayerMovement {
     @Shadow public abstract Player player();
+    @Unique
     private int comboCount;
 
     public ClientPlayerMovementMixin(Player player) {
@@ -24,6 +28,11 @@ public abstract class ClientPlayerMovementMixin extends PlayerMovement {
     @Inject(method = "update", at = @At(value = "HEAD"), remap=false)
     public void update(CallbackInfo ci) {
         calculateMeleeStaminaCost();
+
+        // Cancels the Better Combat attack if the player is out of stamina.
+        if (this.stamina().isDepleted()) {
+            ((MinecraftClient_BetterCombat) Minecraft.getInstance()).cancelUpswing();
+        }
     }
 
     /**